Re-wrote jgmenu_run (again)
jgmenu_run remains a convenience wrapper for jgmenu, but it now has the
following syntax:

	jgmenu_run (<file> | <command>) [<options>]

Any file specified as the first argument must be one of the following:
  - jgmenu type .csv file
  - XDG .menu file

The following 'commands' are available:
    pmenu               Menu based on .directory and .desktop files
    xdg                 Menu based on .menu file
    csv                 Menu based on .csv file
    simple              Simple fall-back menu

Please note, jgmenu-xdg is still at early stages of development, so the
xdg part of this wrapper is merely written in anticipation.

Any 'options' are not used by jgmenu_run, but are passed on to jgmenu.

'cache' will be added to the list of commands later. It is intended
to automate the process of creating icon-cache.

The 'xdg' command searches for .menu files in standard XDG directories.
The 'csv' command only looks for ~/.config/jgmenu/default.csv

$XDG_MENU_PREFIX can be set before running 'jgmenu_run xdg' in order to
specify an *applications.menu file. This is useful if there are several
.menu files (e.g. on a system with several Desktop Environments).