Added theme.c
Two config options were added (both are set to 0 be default)
  - ignore_xsettings
  - ignore_icon_cache

By default, icon-cache will be used if it exists. I.e. the icon-theme
will be set to "jgmenu-cache" and the icons will be read from
~/.local/share/icons/jgmenu-cache/

If ignore_icon_cache=1 OR icon-cache has not be created,
jgmenu will search for an icon theme in the following order:

  - xsettings daemon (unless ignore_xsettings=1)
  - ~/.config/jgmenu/jgmenurc
  - set to "default"