Removed "px" from font-strings
pango-font-description-from-string() interprets the size without
a suffix as "points". If "px" is added, it will be read as pixels.

Using "points" enables consistency with other applications. This is
particularly relevant in config-xs.c

The font can be specified in pixels by just adding "px" to the
font-string in the jgmenurc config file.

Helped-by: Ovidiu