jgmenu.c: re-wrote functions to accept linked list
It has got to a stage where we want jgmenu.c to do far more than it was
originally written for.

Since the first commit, menu items have been read into a simple dynamic
array (=vector). Although the items have been doubly-linked, most of the
key and mouse events have relied on simple pointer arithmetic which
assumes an array.

This commit makes jgmenu.c accept a linked list not stored in a straight
array. This enables features such as filtering. It also prepares for such
things as pipe-menus and alternative menu layouts.

The code now relies on two lists: "menu.master" and "menu.filter".
  - master contains all menu items
  - filter contains the current submenu filtered on anything the user
    has typed as a search pattern

When the menu is open, start typing to activate filter.