Created unity-hack
This is a quick hack to give the Unity launcher an invisible and short-lived
"override_redirect=False" window. This is necessary in order to
  - stop the launcher-button flashing for several seconds;
  - enable the launcher-button to be clicked again.

If jgmenu's "override_redirect" window attribute is set to "True" in
x11-ui.c, a new icon will appear in the launcher which we don't want.

If anyone needs this for anything other than Ubuntu's Unity, I'll consider
re-writing it in C to avoid the python Gtk dependency.