jgmenu.c: created nodal structure
This nodal tree is a layer of abstraction which sits on top of
menu.master (which is the full linked list of menu-items).

Each node knows its parent and has a list of menu-items hanging off it.

Backspace can now be used to checkout the parent node.

jgmenu.c still contains code written for the original dynamic array. This
is a bit untidy and I will gradually re-write all code to refer to the
linked list.