jgmenu.c: added support for "stay_alive" mode
Added config options:
  - stay_alive
  - hide_on_startup

jgmenu will default to the "old" behaviour if these are not defined.

If "stay_alive" is set to 1, the menu will "hide" rather than "exit"
when the following events occur:
  - clicking on menu item
  - clicking outside the menu
  - pressing escape
When in this hidden mode, a USR1 signal will "un-hide" the menu.

Setting "hide_on_startup" to 1, will start jgmenu in a "hidden" state.
This is useful for starting jgmenu during the boot process and then
sending a `killall -SIGUSR1 jgmenu` to show the menu.

We don't check for multiple instances of jgmenu running. If you have
started more than one and then send a USR1 signal, both will try to
"awake" but one will die() as only one can grab the keyboard.  For this
reason, the "stay_alive" mode does not work with pipe-menus as these
are only fake at the minute and actually just invoke another jgmenu.