jgmenu_run: refactor UNIX socket construct
Only send tint2 environment variables to jgmenu's UNIX socket if jgmenu
is running. Otherwise it pollutes the socket with things we do not want.

Check the existance of lockfile to determine if jgmenu is running. pgrep
is slow and we want to keep the 'awake' call is quick as possible.