Align to vertical panel
Without tint2-button environmental variables, jgmenu is aligned to the
left hand edge of a horizontal panel and the top edge of a vertical
panel.

The edge of the panel is simply calculated as a function of panel length
and margin. Although this is adequate for a many cases, it does work in
"compact" mode (panel_shrink=1); nor does it take into account where along
the panel the launcher or execp item is.