jgmenu.c: allow menu.sel == NULL
When the pointer is not over a selectable item, menu.sel is set to NULL
and the 'item' highlighting disappears. If the keyboard is used to
navigate whilst in this state, the last selected item is resumed.

This behaviour is more closely aligned with how most menus work and also
prepares jgmenu for controlling 'widgets' with the pointer.