Add config options for menu height (issue #57)
Add
  - menu_height_min
  - menu_height_max
  - menu_height_mode

menu_height_{min,max} set the min and max height of the root menu. If
these are set to the same, menu will be fixed at that height. If set to
zero, they will be ignored.

menu_height_mode can be either 'static' or 'dynamic'.

'static' is the default and is backward compatible. It means that the
height of the initial root menu will be used for any subsequent ^root().

'dynamic' means that the menu height will be re-calculated every time
the root menu is redefined using ^root().

Single window mode can be achieved with the following configuration:

csv_cmd             = pmenu
menu_height_mode    = dynamic
hide_back_items     = 0
csv_single_window   = 1