apps: support built-in schema
If no schema file exists, the values in src/schema.h will be used.
This creates a default menu which is not dependent on any menu packages
or user/package maintainer effort.

The jgmenu-apps.c + desktop.c + dirs.c code base is much simpler to work
on compared with the lx and pmenu modules.