jgmenu.c: handle unsupported locales
We need to call XSetLocaleModifiers() in order to support utf8 search.

It appears that X does not support all locales, for example en_HK.UTF-8.
In this situation, we gracefully fallback to using LANG=C so that search
can still be supported.

Also, NULL check XDestroyIC() otherwise, we crash on closing sub-menus when
using unsupported locales.