widgets.c: fix cursor-motion bug
In widgets_set_pointer_position() only handle cursor-motion events for
widgets with an action. This is already how cursor-button events are
handled, and it is required to avoid background @rects and @text item
blocking (getting in the way of) the @rect items with action that should
be shown as selected on mouse-over.

Consider the following overlapping widgets where the items are rendered
in the order 1,2,3. We want item 2 (with the action) to respond to
cursor-motion events.

1. background-rect
2. button-rect (with on-click-action)
3. button-text

In the same function we also iterator over the widget items in the reverse
order to which they are drawn, so that in the case of overlapping widgets,
the one drawn on top is the one which also gets receives the cursor-motion
event.

Fixes issue #167