contrib: Fix xfce4-panel registration
Presumably due to header file reorganization,

XFCE_PANEL_PLUGIN_REGISTER(jgmenu_construct);

did not have any effect anymore because the macro is not defined
in <libxfce4panel/xfce-panel-plugin.h> anymore.  Current GCC version
parse this line as a declaration of a function called
XFCE_PANEL_PLUGIN_REGISTER with an implicit int return type
and without a parameter list/prototype, which is the reason why this
did not result in a compilation error.